@@ -1,7 +1,7 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 
 buildscript {
-    ext.kotlin_version = '1.2.61'
+    ext.kotlin_version = '1.3.11'
 
     repositories {
         jcenter()
@@ -11,7 +11,7 @@ buildscript {
         }
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.1.3'
+        classpath 'com.android.tools.build:gradle:3.3.0'
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
 
         // NOTE: Do not place your application dependencies here; they belong
@@ -40,7 +40,7 @@ allprojects {
 }
 
 ext {
-    buildToolsVersion = "28.0.0"
+    buildToolsVersion = "28.0.3"
     minSdkVersion = 21
     compileSdkVersion = 28
     targetSdkVersion = 28
