@@ -27,13 +27,15 @@ const plugins = [
   ],
   '@babel/plugin-transform-sticky-regex',
   '@babel/plugin-transform-unicode-regex',
-  [
-    // TODO: We can get this faster by tweaking with options, but have to test thoroughly...
-    'module:fast-async',
-    {
-      spec: true,
-    },
-  ],
+  // TODO: fast-async is faster and cleaner, but causes a weird issue on older Android RN targets without jsc-android
+  '@babel/plugin-transform-async-to-generator',
+  // [
+  //   // TODO: We can get this faster by tweaking with options, but have to test thoroughly...
+  //   'module:fast-async',
+  //   {
+  //     spec: true,
+  //   },
+  // ],
 ]
 
 module.exports = {
@@ -53,7 +55,7 @@ module.exports = {
       ],
     },
     test: {
-      plugins,
+      plugins: [...plugins, '@babel/plugin-syntax-jsx'],
     },
   },
 }
