@@ -1,5 +1,6 @@
 import React from 'react'
 import { compose, withStateHandlers, withHandlers } from 'recompose'
+import { Q } from '@nozbe/watermelondb'
 import withObservables from '@nozbe/with-observables'
 
 import ListItem from 'components/ListItem'
@@ -39,11 +40,10 @@ const BlogList = ({ blogs, setActiveItem, activeItem }) => (
 )
 
 const enhance = compose(
-  withObservables([], ({ database }) => ({
+  withObservables(['search'], ({ database, search }) => ({
     blogs: database.collections
       .get('blogs')
-      .query()
-      .observe(),
+      .query(Q.where('name', Q.like(`%${Q.sanitizeLikeString(search)}%`))),
   })),
   withStateHandlers(
     {
