@@ -7,14 +7,19 @@ import { values } from 'rambdax'
 
 import { invariant } from '../utils/common'
 
-import CollectionMap from '../CollectionMap'
-
 import type { DatabaseAdapter } from '../adapters/type'
 import type Model from '../Model'
 import type { CollectionChangeSet } from '../Collection'
 import type { TableName, AppSchema } from '../Schema'
 
-// Database is the owner of all Collections and the DatabaseAdapter
+import CollectionMap from './CollectionMap'
+import ActionQueue, { type ActionInterface } from './ActionQueue'
+
+type DatabaseProps = $Exact<{
+  adapter: DatabaseAdapter,
+  modelClasses: Array<Class<Model>>,
+  actionsEnabled?: boolean,
+}>
 
 export default class Database {
   adapter: DatabaseAdapter
@@ -23,21 +28,24 @@ export default class Database {
 
   collections: CollectionMap
 
-  constructor({
-    adapter,
-    modelClasses,
-  }: $Exact<{
-    adapter: DatabaseAdapter,
-    modelClasses: Array<Class<Model>>,
-  }>): void {
+  _actionQueue = new ActionQueue()
+
+  _actionsEnabled: boolean
+
+  constructor({ adapter, modelClasses, actionsEnabled = false }: DatabaseProps): void {
     this.adapter = adapter
     this.schema = adapter.schema
     this.collections = new CollectionMap(this, modelClasses)
+    this._actionsEnabled = actionsEnabled
   }
 
   // Executes multiple prepared operations
   // (made with `collection.prepareCreate` and `record.prepareUpdate`)
   async batch(...records: $ReadOnlyArray<Model>): Promise<void> {
+    this._ensureInAction(
+      `Database.batch() can only be called from inside of an Action. See docs for more details.`,
+    )
+
     const operations = records.map(record => {
       invariant(
         !record._isCommitted || record._hasPendingUpdate,
@@ -63,6 +71,11 @@ export default class Database {
     })
   }
 
+  // TODO: Document me!
+  action<T>(work: ActionInterface => Promise<T>, description?: string): Promise<T> {
+    return this._actionQueue.enqueue(work, description)
+  }
+
   // Emits a signal immediately, and on change in any of the passed tables
   withChangesForTables(tables: TableName<any>[]): Observable<CollectionChangeSet<any> | null> {
     const changesSignals = tables.map(table => this.collections.get(table).changes)
@@ -80,5 +93,9 @@ export default class Database {
     values(this.collections.map).forEach(collection => {
       collection.unsafeClearCache()
     })
+  }
+
+  _ensureInAction(error: string): void {
+    this._actionsEnabled && invariant(this._actionQueue.isRunning, error)
   }
 }
