@@ -20,10 +20,12 @@ import cond from '../utils/fp/cond'
 import partition from '../utils/fp/partition'
 import isObject from '../utils/fp/isObject'
 import invariant from '../utils/common/invariant'
+import type { $RE } from '../types'
 
 import { type TableName, type ColumnName, columnName } from '../Schema'
 
 export type NonNullValue = number | string | boolean
+export type NonNullValues = number[] | string[] | boolean[]
 export type Value = NonNullValue | null
 export type CompoundValue = Value | Value[]
 
@@ -38,15 +40,16 @@ export type Operator =
   | 'oneOf'
   | 'notIn'
   | 'between'
+  | 'like'
 
-export type ColumnDescription = $Exact<{ column: ColumnName }>
+export type ColumnDescription = $RE<{ column: ColumnName }>
 export type ComparisonRight =
-  | $Exact<{ value: Value }>
-  | $Exact<{ values: NonNullValue[] }>
+  | $RE<{ value: Value }>
+  | $RE<{ values: NonNullValues }>
   | ColumnDescription
-export type Comparison = $Exact<{ operator: Operator, right: ComparisonRight }>
+export type Comparison = $RE<{ operator: Operator, right: ComparisonRight }>
 
-export type WhereDescription = $Exact<{
+export type WhereDescription = $RE<{
   type: 'where',
   left: ColumnName,
   comparison: Comparison,
@@ -54,16 +57,16 @@ export type WhereDescription = $Exact<{
 
 /* eslint-disable-next-line */
 export type Where = WhereDescription | And | Or
-export type And = $Exact<{ type: 'and', conditions: Where[] }>
-export type Or = $Exact<{ type: 'or', conditions: Where[] }>
-export type On = $Exact<{
+export type And = $RE<{ type: 'and', conditions: Where[] }>
+export type Or = $RE<{ type: 'or', conditions: Where[] }>
+export type On = $RE<{
   type: 'on',
   table: TableName<any>,
   left: ColumnName,
   comparison: Comparison,
 }>
 export type Condition = Where | On
-export type QueryDescription = $Exact<{ where: Where[], join: On[] }>
+export type QueryDescription = $RE<{ where: Where[], join: On[] }>
 
 // Note: These operators are designed to match SQLite semantics
 // to ensure that iOS, Android, web, and Query observation yield exactly the same results
@@ -145,7 +148,11 @@ export function lte(valueOrColumn: NonNullValue | ColumnDescription): Comparison
 // Value in a set (SQLite IN semantics)
 // Note:
 // - `null` in `values` is not allowed!
-export function oneOf(values: NonNullValue[]): Comparison {
+export function oneOf(values: NonNullValues): Comparison {
+  if (process.env.NODE_ENV !== 'production') {
+    invariant(Array.isArray(values), `argument passed to oneOf() is not an array`)
+  }
+
   return { operator: 'oneOf', right: { values } }
 }
 
@@ -153,13 +160,26 @@ export function oneOf(values: NonNullValue[]): Comparison {
 // Note:
 // - `null` in `values` is not allowed!
 // - (null NOT IN (1, 2, 3)) == false
-export function notIn(values: NonNullValue[]): Comparison {
+export function notIn(values: NonNullValues): Comparison {
+  if (process.env.NODE_ENV !== 'production') {
+    invariant(Array.isArray(values), `argument passed to notIn() is not an array`)
+  }
+
   return { operator: 'notIn', right: { values } }
 }
 
 // Number is between two numbers (greater than or equal left, and less than or equal right)
 export function between(left: number, right: number): Comparison {
-  return { operator: 'between', right: { values: [left, right] } }
+  const values: number[] = [left, right]
+  return { operator: 'between', right: { values } }
+}
+
+export function like(value: string): Comparison {
+  return { operator: 'like', right: { value } }
+}
+
+export function sanitizeLikeString(value: string): string {
+  return value.replace(/[^a-zA-Z0-9]/g, '_')
 }
 
 export function column(name: ColumnName): ColumnDescription {
