@@ -2,10 +2,10 @@ import { MockTask, MockProject, mockDatabase } from '../__tests__/testModels'
 
 import Relation from './index'
 
-describe('watermelondb/Relation', () => {
+describe('Relation', () => {
   it('gets id', () => {
-    const { tasksCollection } = mockDatabase()
-    const primary = new MockTask(tasksCollection, { project_id: 's1' })
+    const { tasks } = mockDatabase()
+    const primary = new MockTask(tasks, { project_id: 's1' })
 
     const relation = new Relation(primary, 'mock_projects', 'project_id', { isImmutable: false })
     expect(relation.id).toBe('s1')
@@ -14,12 +14,11 @@ describe('watermelondb/Relation', () => {
     primary.projectId = 's2'
     expect(relation.id).toBe('s2')
   })
-
   it('sets id', () => {
-    const { tasksCollection, projectsCollection } = mockDatabase()
+    const { tasks, projects } = mockDatabase()
 
-    const primary = new MockTask(tasksCollection, { project_id: null })
-    const secondary = new MockProject(projectsCollection, {
+    const primary = new MockTask(tasks, { project_id: null })
+    const secondary = new MockProject(projects, {
       id: 's1',
     })
 
@@ -31,11 +30,10 @@ describe('watermelondb/Relation', () => {
     expect(relation.id).toBe('s1')
     expect(primary.projectId).toBe('s1')
   })
-
   it('sets record', () => {
-    const { tasksCollection, projectsCollection } = mockDatabase()
-    const primary = new MockTask(tasksCollection, {})
-    const secondary = new MockProject(projectsCollection, { id: 's1' })
+    const { tasks, projects } = mockDatabase()
+    const primary = new MockTask(tasks, {})
+    const secondary = new MockProject(projects, { id: 's1' })
 
     const relation = new Relation(primary, 'mock_projects', 'project_id', { isImmutable: false })
     primary._isEditing = true
@@ -43,15 +41,14 @@ describe('watermelondb/Relation', () => {
 
     expect(relation.id).toBe('s1')
   })
-
   it('allows setting id/record only on create/prepareCreate when immutable', async () => {
-    const { tasksCollection, commentsCollection } = mockDatabase()
+    const { tasks, comments } = mockDatabase()
 
-    const secondary = await tasksCollection.create(mock => {
+    const secondary = await tasks.create(mock => {
       mock.name = 'foo'
     })
 
-    const primary = await commentsCollection.create(mock => {
+    const primary = await comments.create(mock => {
       mock.task.id = secondary.id
     })
 
@@ -63,11 +60,11 @@ describe('watermelondb/Relation', () => {
       }),
     ).toThrow()
 
-    const secondary2 = await commentsCollection.create(mock => {
+    const secondary2 = await comments.create(mock => {
       mock.name = 'bar'
     })
 
-    const primary2 = await commentsCollection.prepareCreate(mock => {
+    const primary2 = await comments.prepareCreate(mock => {
       mock.task.id = secondary.id
       expect(mock.task.id).toBe(secondary.id)
       mock.task.set(secondary2)
@@ -76,15 +73,14 @@ describe('watermelondb/Relation', () => {
 
     expect(primary2.task.id).toBe(secondary2.id)
   })
-
   it('observers related record', async () => {
-    const { tasksCollection, projectsCollection } = mockDatabase()
+    const { tasks, projects } = mockDatabase()
 
-    const secondary = await projectsCollection.create(mock => {
+    const secondary = await projects.create(mock => {
       mock.name = 'foo'
     })
 
-    const primary = await tasksCollection.create(mock => {
+    const primary = await tasks.create(mock => {
       mock.projectId = secondary.id
     })
 
@@ -95,7 +91,7 @@ describe('watermelondb/Relation', () => {
 
     await new Promise(process.nextTick) // give time to propagate
 
-    expect(observer).toBeCalledWith(secondary)
+    expect(observer).toHaveBeenCalledWith(secondary)
 
     await secondary.update(mock => {
       mock.name = 'bar'
@@ -104,15 +100,14 @@ describe('watermelondb/Relation', () => {
     expect(observer).toHaveBeenCalledTimes(2)
     subscription.unsubscribe()
   })
-
   it('fetches current record', async () => {
-    const { tasksCollection, projectsCollection } = mockDatabase()
+    const { tasks, projects } = mockDatabase()
 
-    const secondary = await projectsCollection.create(mock => {
+    const secondary = await projects.create(mock => {
       mock.name = 'foo'
     })
 
-    const primary = await tasksCollection.create(mock => {
+    const primary = await tasks.create(mock => {
       mock.projectId = secondary.id
     })
 
@@ -121,7 +116,7 @@ describe('watermelondb/Relation', () => {
     let currentRecord = await relation.fetch()
     expect(currentRecord).toBe(secondary)
 
-    const newSecondary = await projectsCollection.create(mock => {
+    const newSecondary = await projects.create(mock => {
       mock.name = 'bar'
     })
 
@@ -132,21 +127,9 @@ describe('watermelondb/Relation', () => {
     currentRecord = await relation.fetch()
     expect(currentRecord).toBe(newSecondary)
   })
-
   it('caches observable', () => {
-    const { tasksCollection } = mockDatabase()
-    const model = new MockTask(tasksCollection, {})
-    const relation = new Relation(model, 't1', 'c1', { isImmutable: false })
-
-    const observable1 = relation.observe()
-    const observable2 = relation.observe()
-
-    expect(observable1).toBe(observable2)
-  })
-
-  it('caches observable', () => {
-    const { tasksCollection } = mockDatabase()
-    const model = new MockTask(tasksCollection, {})
+    const { tasks } = mockDatabase()
+    const model = new MockTask(tasks, {})
     const relation = new Relation(model, 't1', 'c1', { isImmutable: false })
 
     const observable1 = relation.observe()
