@@ -2,8 +2,10 @@
 
 import { contains } from 'rambdax'
 
+import logger from '../utils/common/logger'
 import isDevelopment from '../utils/common/isDevelopment'
 import invariant from '../utils/common/invariant'
+import type { $RE } from '../types'
 
 import type Model from '../Model'
 
@@ -11,7 +13,7 @@ export opaque type TableName<+T: Model>: string = string
 export opaque type ColumnName: string = string
 
 export type ColumnType = 'string' | 'number' | 'boolean'
-export type ColumnSchema = $Exact<{
+export type ColumnSchema = $RE<{
   name: ColumnName,
   type: ColumnType,
   isOptional?: boolean,
@@ -22,13 +24,13 @@ export type ColumnMap = { [name: ColumnName]: ColumnSchema }
 
 export type TableSchemaSpec = $Exact<{ name: TableName<any>, columns: ColumnSchema[] }>
 
-export type TableSchema = $Exact<{ name: TableName<any>, columns: ColumnMap }>
+export type TableSchema = $RE<{ name: TableName<any>, columns: ColumnMap }>
 
 type TableMap = { [name: TableName<any>]: TableSchema }
 
 export type SchemaVersion = number
 
-export type AppSchema = $Exact<{ version: SchemaVersion, tables: TableMap }>
+export type AppSchema = $RE<{ version: SchemaVersion, tables: TableMap }>
 
 export function tableName<T: Model>(name: string): TableName<T> {
   return name
@@ -59,13 +61,19 @@ export function validateColumnSchema(column: ColumnSchema): void {
       `Invalid type ${column.type} for column ${column.name} (valid: string, boolean, number)`,
     )
     invariant(
-      !contains(column.name, ['id', 'last_modified', '_changed', '_status']),
-      `You must not define columns with name ${column.name}`,
+      !contains(column.name, ['id', '_changed', '_status']),
+      `You must not define a column with name ${column.name}`,
     )
     if (column.name === 'created_at' || column.name === 'updated_at') {
       invariant(
         column.type === 'number' && !column.isOptional,
         `${column.name} must be of type number and not optional`,
+      )
+    }
+    if (column.name === 'last_modified') {
+      invariant(
+        column.type === 'number',
+        `For compatibility reasons, column last_modified must be of type 'number', and should be optional`,
       )
     }
   }
@@ -78,8 +86,7 @@ export function tableSchema({ name, columns: columnList }: TableSchemaSpec): Tab
     if (column.type === 'bool') {
       column.type = 'boolean'
       if (isDevelopment) {
-        // eslint-disable-next-line
-        console.warn(
+        logger.warn(
           `[DEPRECATION] Column type 'bool' is deprecated — change to 'boolean' (in ${JSON.stringify(
             column,
           )})`,
