@@ -3,27 +3,36 @@ import { stepsForMigration } from './helpers'
 
 describe('schemaMigrations()', () => {
   it('returns a basic schema migrations spec', () => {
-    const migrations = schemaMigrations({
-      minimumVersion: 1,
-      currentVersion: 1,
-      migrations: [],
-    })
+    const migrations = schemaMigrations({ migrations: [] })
     expect(migrations).toEqual({
-      minimumVersion: 1,
-      currentVersion: 1,
-      migrations: [],
+      sortedMigrations: [],
       validated: true,
+      minVersion: 1,
+      maxVersion: 1,
+    })
+
+    const migrations2 = schemaMigrations({ migrations: [{ toVersion: 2, steps: [] }] })
+    expect(migrations2).toEqual({
+      validated: true,
+      minVersion: 1,
+      maxVersion: 2,
+      sortedMigrations: [{ toVersion: 2, steps: [] }],
+    })
+
+    const migrations3 = schemaMigrations({ migrations: [{ toVersion: 4, steps: [] }] })
+    expect(migrations3).toEqual({
+      validated: true,
+      minVersion: 3,
+      maxVersion: 4,
+      sortedMigrations: [{ toVersion: 4, steps: [] }],
     })
   })
   it('returns a complex schema migrations spec', () => {
     const migrations = schemaMigrations({
-      minimumVersion: 1,
-      currentVersion: 5,
       migrations: [
-        { from: 4, to: 5, steps: [] },
+        { toVersion: 4, steps: [] },
         {
-          from: 2,
-          to: 4,
+          toVersion: 3,
           steps: [
             createTable({
               name: 'comments',
@@ -39,8 +48,7 @@ describe('schemaMigrations()', () => {
           ],
         },
         {
-          from: 1,
-          to: 2,
+          toVersion: 2,
           steps: [
             addColumns({
               table: 'posts',
@@ -55,13 +63,24 @@ describe('schemaMigrations()', () => {
     })
     expect(migrations).toEqual({
       validated: true,
-      minimumVersion: 1,
-      currentVersion: 5,
-      migrations: [
-        { from: 4, to: 5, steps: [] },
+      minVersion: 1,
+      maxVersion: 4,
+      sortedMigrations: [
         {
-          from: 2,
-          to: 4,
+          toVersion: 2,
+          steps: [
+            {
+              type: 'add_columns',
+              table: 'posts',
+              columns: [
+                { name: 'subtitle', type: 'string', isOptional: true },
+                { name: 'is_pinned', type: 'boolean' },
+              ],
+            },
+          ],
+        },
+        {
+          toVersion: 3,
           steps: [
             {
               type: 'create_table',
@@ -78,94 +97,59 @@ describe('schemaMigrations()', () => {
             },
           ],
         },
-        {
-          from: 1,
-          to: 2,
-          steps: [
-            {
-              type: 'add_columns',
-              table: 'posts',
-              columns: [
-                { name: 'subtitle', type: 'string', isOptional: true },
-                { name: 'is_pinned', type: 'boolean' },
-              ],
-            },
-          ],
-        },
+        { toVersion: 4, steps: [] },
       ],
     })
   })
   it('throws if migration spec is malformed', () => {
+    expect(() => schemaMigrations({ migrations: [{}] })).toThrow(/Invalid migration/)
+    expect(() => schemaMigrations({ migrations: [{ toVersion: 0, steps: [] }] })).toThrow(
+      /minimum.*is 2/i,
+    )
+    expect(() => schemaMigrations({ migrations: [{ toVersion: 1, steps: [] }] })).toThrow(
+      /minimum.*is 2/i,
+    )
     expect(() =>
       schemaMigrations({
-        currentVersion: 1,
-        migrations: [],
-      }),
-    ).toThrow(/minimum schema version/i)
-    expect(() =>
-      schemaMigrations({
-        minimumVersion: 1,
-        migrations: [],
-      }),
-    ).toThrow(/current schema version/i)
-    expect(() =>
-      schemaMigrations({
-        minimumVersion: 0,
-        currentVersion: 1,
-        migrations: [],
-      }),
-    ).toThrow(/at least 1/)
-    expect(() =>
-      schemaMigrations({
-        minimumVersion: 2,
-        currentVersion: 1,
-        migrations: [],
-      }),
-    ).toThrow(/greater than/)
-    expect(() =>
-      schemaMigrations({
-        minimumVersion: 1,
-        currentVersion: 1,
-        migrations: [{}],
-      }),
-    ).toThrow(/Invalid migration/)
-    expect(() =>
-      schemaMigrations({
-        minimumVersion: 1,
-        currentVersion: 1,
-        migrations: [{ from: 1, to: 1 }],
-      }),
-    ).toThrow(/greater than/)
-    expect(() =>
-      schemaMigrations({
-        minimumVersion: 1,
-        currentVersion: 2,
-        migrations: [{ from: 1, to: 2, steps: [{ table: 'x' }] }],
+        migrations: [{ toVersion: 2, steps: [{ table: 'x' }] }],
       }),
     ).toThrow(/Invalid migration steps/)
   })
-  it(`throws if migrations don't cover the whole migrable range`, () => {
+  it(`throws if there are gaps or duplicates in migrations`, () => {
+    expect(() =>
+      schemaMigrations({ migrations: [{ toVersion: 2, steps: [] }, { toVersion: 2, steps: [] }] }),
+    ).toThrow(/duplicates/)
     expect(() =>
       schemaMigrations({
-        minimumVersion: 1,
-        currentVersion: 3,
-        migrations: [{ from: 2, to: 3, steps: [] }],
+        migrations: [
+          { toVersion: 5, steps: [] },
+          { toVersion: 4, steps: [] },
+          { toVersion: 2, steps: [] },
+        ],
       }),
-    ).toThrow(/covers schema versions/)
+    ).toThrow(/gaps/)
+
+    // missing migrations from 2 to x are ok
     expect(() =>
       schemaMigrations({
-        minimumVersion: 1,
-        currentVersion: 3,
-        migrations: [{ from: 1, to: 2, steps: [] }],
+        migrations: [
+          { toVersion: 6, steps: [] },
+          { toVersion: 5, steps: [] },
+          { toVersion: 4, steps: [] },
+        ],
       }),
-    ).toThrow(/cover schema versions/)
+    ).not.toThrow()
+
+    // chronological is ok too
     expect(() =>
       schemaMigrations({
-        minimumVersion: 1,
-        currentVersion: 4,
-        migrations: [{ from: 1, to: 2, steps: [] }, { from: 2, to: 3, steps: [] }],
+        migrations: [
+          { toVersion: 4, steps: [] },
+          { toVersion: 5, steps: [] },
+          { toVersion: 6, steps: [] },
+        ],
       }),
-    ).toThrow(/covers schema versions/)
+    ).not.toThrow()
   })
 })
 
@@ -210,24 +194,36 @@ describe('migration execution helpers', () => {
     })
 
     const migrations = schemaMigrations({
-      minimumVersion: 1,
-      currentVersion: 5,
       migrations: [
-        { from: 4, to: 5, steps: [step2, step3] },
-        { from: 2, to: 4, steps: [] },
-        { from: 1, to: 2, steps: [step1] },
+        { toVersion: 5, steps: [step2, step3] },
+        { toVersion: 4, steps: [] },
+        { toVersion: 3, steps: [step1] },
       ],
     })
 
-    expect(stepsForMigration({ migrations, fromVersion: 1, toVersion: 2 })).toEqual([step1])
-    expect(stepsForMigration({ migrations, fromVersion: 1, toVersion: 4 })).toEqual([step1])
-    expect(stepsForMigration({ migrations, fromVersion: 1, toVersion: 5 })).toEqual([
+    expect(stepsForMigration({ migrations, fromVersion: 2, toVersion: 3 })).toEqual([step1])
+    expect(stepsForMigration({ migrations, fromVersion: 2, toVersion: 4 })).toEqual([step1])
+    expect(stepsForMigration({ migrations, fromVersion: 2, toVersion: 5 })).toEqual([
       step1,
       step2,
       step3,
     ])
-    expect(stepsForMigration({ migrations, fromVersion: 2, toVersion: 5 })).toEqual([step2, step3])
-    expect(stepsForMigration({ migrations, fromVersion: 2, toVersion: 4 })).toEqual([])
+    expect(stepsForMigration({ migrations, fromVersion: 3, toVersion: 5 })).toEqual([step2, step3])
+    expect(stepsForMigration({ migrations, fromVersion: 3, toVersion: 4 })).toEqual([])
     expect(stepsForMigration({ migrations, fromVersion: 4, toVersion: 5 })).toEqual([step2, step3])
+
+    // if no available steps, return null
+    expect(
+      stepsForMigration({
+        migrations: schemaMigrations({ migrations: [] }),
+        fromVersion: 1,
+        toVersion: 2,
+      }),
+    ).toEqual(null)
+    expect(stepsForMigration({ migrations, fromVersion: 1, toVersion: 2 })).toEqual(null)
+    expect(stepsForMigration({ migrations, fromVersion: 1, toVersion: 3 })).toEqual(null)
+    expect(stepsForMigration({ migrations, fromVersion: 1, toVersion: 5 })).toEqual(null)
+    expect(stepsForMigration({ migrations, fromVersion: 3, toVersion: 6 })).toEqual(null)
+    expect(stepsForMigration({ migrations, fromVersion: 5, toVersion: 6 })).toEqual(null)
   })
 })
