@@ -10,6 +10,7 @@ import type {
   SchemaMigrations,
   CreateTableMigrationStep,
   AddColumnsMigrationStep,
+  MigrationStep,
 } from '../../../Schema/migrations'
 import { stepsForMigration } from '../../../Schema/migrations/helpers'
 import type { SerializedQuery } from '../../../Query'
@@ -40,7 +41,7 @@ export default class LokiExecutor {
 
   _testLokiAdapter: ?LokiMemoryAdapter
 
-  cachedRecords: Set<RecordId> = new Set([])
+  cachedRecords: Map<TableName<any>, Set<RecordId>> = new Map()
 
   constructor(options: LokiExecutorOptions): void {
     const { dbName, schema, migrationsExperimental, _testLokiAdapter } = options
@@ -55,8 +56,29 @@ export default class LokiExecutor {
     await this._migrateIfNeeded()
   }
 
+  isCached(table: TableName<any>, id: RecordId): boolean {
+    const cachedSet = this.cachedRecords.get(table)
+    return cachedSet ? cachedSet.has(id) : false
+  }
+
+  markAsCached(table: TableName<any>, id: RecordId): void {
+    const cachedSet = this.cachedRecords.get(table)
+    if (cachedSet) {
+      cachedSet.add(id)
+    } else {
+      this.cachedRecords.set(table, new Set([id]))
+    }
+  }
+
+  removeFromCache(table: TableName<any>, id: RecordId): void {
+    const cachedSet = this.cachedRecords.get(table)
+    if (cachedSet) {
+      cachedSet.delete(id)
+    }
+  }
+
   find(table: TableName<any>, id: RecordId): CachedFindResult {
-    if (this.cachedRecords.has(id)) {
+    if (this.isCached(table, id)) {
       return id
     }
 
@@ -66,7 +88,7 @@ export default class LokiExecutor {
       return null
     }
 
-    this.cachedRecords.add(id)
+    this.markAsCached(table, id)
     return sanitizedRaw(raw, this.schema.tables[table])
   }
 
@@ -81,7 +103,7 @@ export default class LokiExecutor {
 
   create(table: TableName<any>, raw: RawRecord): void {
     this.loki.getCollection(table).insert(raw)
-    this.cachedRecords.add(raw.id)
+    this.markAsCached(table, raw.id)
   }
 
   update(table: TableName<any>, rawRecord: RawRecord): void {
@@ -97,7 +119,7 @@ export default class LokiExecutor {
     const collection = this.loki.getCollection(table)
     const record = collection.by('id', id)
     collection.remove(record)
-    this.cachedRecords.delete(id)
+    this.removeFromCache(table, id)
   }
 
   markAsDeleted(table: TableName<any>, id: RecordId): void {
@@ -106,7 +128,7 @@ export default class LokiExecutor {
     if (record) {
       record._status = 'deleted'
       collection.update(record)
-      this.cachedRecords.delete(id)
+      this.removeFromCache(table, id)
     }
   }
 
@@ -252,13 +274,20 @@ export default class LokiExecutor {
       await this.unsafeResetDatabase()
     } else if (dbVersion > 0 && dbVersion < schemaVersion) {
       logger.log('[DB][Worker] Database has old schema version. Migration is required.')
-      const { migrations } = this
-      if (migrations) {
-        logger.log('[DB][Worker] Migrations available, migrating schema…')
-        this._migrate(migrations, dbVersion)
+      const migrationSteps = this._getMigrationSteps(dbVersion)
+
+      if (migrationSteps) {
+        logger.log(`[DB][Worker] Migrating from version ${dbVersion} to ${this.schema.version}...`)
+        try {
+          await this._migrate(migrationSteps)
+        } catch (error) {
+          logger.error('[DB][Worker] Migration failed', error)
+          throw error
+        }
       } else {
-        // TODO: Delete this altogether? Or put under "development" flag only?
-        logger.warn('[DB][Worker] No migrations available, resetting database')
+        logger.warn(
+          '[DB][Worker] Migrations not available for this version range, resetting database instead',
+        )
         await this.unsafeResetDatabase()
       }
     } else {
@@ -267,14 +296,22 @@ export default class LokiExecutor {
     }
   }
 
-  _migrate(migrations: SchemaMigrations, fromVersion: SchemaVersion): void {
-    const migrationSteps = stepsForMigration({
+  _getMigrationSteps(fromVersion: SchemaVersion): ?(MigrationStep[]) {
+    // TODO: Remove this after migrations are shipped
+    const { migrations } = this
+    if (!migrations) {
+      return null
+    }
+
+    return stepsForMigration({
       migrations,
       fromVersion,
       toVersion: this.schema.version,
     })
+  }
 
-    migrationSteps.forEach(step => {
+  async _migrate(steps: MigrationStep[]): Promise<void> {
+    steps.forEach(step => {
       if (step.type === 'create_table') {
         this._executeCreateTableMigration(step)
       } else if (step.type === 'add_columns') {
@@ -287,7 +324,7 @@ export default class LokiExecutor {
     // Set database version
     this._databaseVersion = this.schema.version
 
-    logger.log('[DB][Worker] Database migrations performed')
+    logger.log(`[DB][Worker] Migration successful`)
   }
 
   _executeCreateTableMigration({ name, columns }: CreateTableMigrationStep): void {
@@ -317,11 +354,11 @@ export default class LokiExecutor {
     return records.map(raw => {
       const { id } = raw
 
-      if (this.cachedRecords.has(id)) {
+      if (this.isCached(table, id)) {
         return id
       }
 
-      this.cachedRecords.add(id)
+      this.markAsCached(table, id)
       return sanitizedRaw(raw, this.schema.tables[table])
     })
   }
