@@ -1,6 +1,6 @@
 import encodeName from './index'
 
-describe('watermelondb/adapters/sqlite/encodeName', () => {
+describe('SQLite encodeName', () => {
   it('encodes names', () => {
     expect(encodeName(`from`)).toBe(`"from"`)
     expect(encodeName(`tasks`)).toBe(`"tasks"`)
