@@ -24,11 +24,11 @@ describe('encodeSchema', () => {
     })
 
     const expectedSchema =
-      'create table "tasks" ("id" primary key, "_changed", "_status", "last_modified", "author_id", "order", "created_at");' +
+      'create table "tasks" ("id" primary key, "_changed", "_status", "author_id", "order", "created_at");' +
       'create index tasks_author_id on "tasks" ("author_id");' +
       'create index tasks_order on "tasks" ("order");' +
       'create index tasks__status on "tasks" ("_status");' +
-      'create table "comments" ("id" primary key, "_changed", "_status", "last_modified", "is_ended", "reactions");' +
+      'create table "comments" ("id" primary key, "_changed", "_status", "is_ended", "reactions");' +
       'create index comments__status on "comments" ("_status");'
 
     expect(encodeSchema(testSchema)).toBe(expectedSchema)
@@ -58,7 +58,7 @@ describe('encodeSchema', () => {
     const expectedSQL =
       `alter table "posts" add "subtitle";` +
       `update "posts" set "subtitle" = null;` +
-      `create table "comments" ("id" primary key, "_changed", "_status", "last_modified", "post_id", "body");` +
+      `create table "comments" ("id" primary key, "_changed", "_status", "post_id", "body");` +
       `create index comments_post_id on "comments" ("post_id");` +
       `create index comments__status on "comments" ("_status");` +
       `alter table "posts" add "author_id";` +
