@@ -13,7 +13,7 @@ import {
 import type Model, { RecordId } from '../../Model'
 import type Query from '../../Query'
 import type { TableName, AppSchema, SchemaVersion } from '../../Schema'
-import type { SchemaMigrations } from '../../Schema/migrations'
+import type { SchemaMigrations, MigrationStep } from '../../Schema/migrations'
 import type { DatabaseAdapter, CachedQueryResult, CachedFindResult, BatchOperation } from '../type'
 import {
   type DirtyFindResult,
@@ -37,8 +37,8 @@ export type SQLiteArg = string | boolean | number | null
 export type SQLiteQuery = [SQL, SQLiteArg[]]
 
 type NativeBridgeBatchOperation =
-  | ['execute', SQL, SQLiteArg[]]
-  | ['create', RecordId, SQL, SQLiteArg[]]
+  | ['execute', TableName<any>, SQL, SQLiteArg[]]
+  | ['create', TableName<any>, RecordId, SQL, SQLiteArg[]]
   | ['markAsDeleted', TableName<any>, RecordId]
   | ['destroyPermanently', TableName<any>, RecordId]
 // | ['setLocal', string, string]
@@ -53,7 +53,7 @@ type NativeBridgeType = {
   setUpWithSchema: (ConnectionTag, string, SQL, SchemaVersion) => Promise<void>,
   setUpWithMigrations: (ConnectionTag, string, SQL, SchemaVersion, SchemaVersion) => Promise<void>,
   find: (ConnectionTag, TableName<any>, RecordId) => Promise<DirtyFindResult>,
-  query: (ConnectionTag, SQL) => Promise<DirtyQueryResult>,
+  query: (ConnectionTag, TableName<any>, SQL) => Promise<DirtyQueryResult>,
   count: (ConnectionTag, SQL) => Promise<number>,
   batch: (ConnectionTag, NativeBridgeBatchOperation[]) => Promise<void>,
   getDeletedRecords: (ConnectionTag, TableName<any>) => Promise<RecordId[]>,
@@ -114,40 +114,53 @@ export default class SQLiteAdapter implements DatabaseAdapter {
     const status = await Native.initialize(this._tag, this._dbName, this.schema.version)
 
     if (status.code === 'schema_needed') {
-      logger.log('[DB] Database needs setup. Setting up schema.')
       await this._setUpWithSchema()
     } else if (status.code === 'migrations_needed') {
-      logger.log('[DB] Database needs migrations')
-      const { databaseVersion } = status
-      invariant(databaseVersion > 0, 'Invalid database schema version')
-
-      if (this.migrations) {
-        const migrationSQL = this._encodedMigrations(this.migrations, databaseVersion)
-        await Native.setUpWithMigrations(
-          this._tag,
-          this._dbName,
-          migrationSQL,
-          databaseVersion,
-          this.schema.version,
-        )
-        logger.log('[DB] Migrations applied successfully')
-      } else {
-        // TODO: Temporary, remove this branch later
-        logger.warn('[DB] Migrations not available. Resetting database instead')
-        await this._setUpWithSchema()
-      }
+      await this._setUpWithMigrations(status.databaseVersion)
     } else {
       invariant(status.code === 'ok', 'Invalid database initialization status')
     }
   }
 
+  async _setUpWithMigrations(databaseVersion: SchemaVersion): Promise<void> {
+    logger.log('[DB] Database needs migrations')
+    invariant(databaseVersion > 0, 'Invalid database schema version')
+
+    const migrationSteps = this._migrationSteps(databaseVersion)
+
+    if (migrationSteps) {
+      logger.log(`[DB] Migrating from version ${databaseVersion} to ${this.schema.version}...`)
+
+      try {
+        await Native.setUpWithMigrations(
+          this._tag,
+          this._dbName,
+          this._encodeMigrations(migrationSteps),
+          databaseVersion,
+          this.schema.version,
+        )
+        logger.log('[DB] Migration successful')
+      } catch (error) {
+        logger.error('[DB] Migration failed', error)
+        throw error
+      }
+    } else {
+      logger.warn(
+        '[DB] Migrations not available for this version range, resetting database instead',
+      )
+      await this._setUpWithSchema()
+    }
+  }
+
   async _setUpWithSchema(): Promise<void> {
-    return Native.setUpWithSchema(
+    logger.log(`[DB] Setting up database with schema version ${this.schema.version}`)
+    await Native.setUpWithSchema(
       this._tag,
       this._dbName,
       this._encodedSchema(),
       this.schema.version,
     )
+    logger.log(`[DB] Schema set up successfully`)
   }
 
   async find(table: TableName<any>, id: RecordId): Promise<CachedFindResult> {
@@ -163,7 +176,7 @@ export default class SQLiteAdapter implements DatabaseAdapter {
     return devLogQuery(
       async () =>
         sanitizeQueryResult(
-          await Native.query(this._tag, encodeQuery(query)),
+          await Native.query(this._tag, query.table, encodeQuery(query)),
           this.schema.tables[query.table],
         ),
       query,
@@ -181,14 +194,14 @@ export default class SQLiteAdapter implements DatabaseAdapter {
         operations.map(([type, record]) => {
           switch (type) {
             case 'create':
-              return ['create', record.id, ...encodeInsert(record)]
+              return ['create', record.table, record.id, ...encodeInsert(record)]
             case 'markAsDeleted':
               return ['markAsDeleted', record.table, record.id]
             case 'destroyPermanently':
               return ['destroyPermanently', record.table, record.id]
             default:
               // case 'update':
-              return ['execute', ...encodeUpdate(record)]
+              return ['execute', record.table, ...encodeUpdate(record)]
           }
         }),
       )
@@ -225,14 +238,22 @@ export default class SQLiteAdapter implements DatabaseAdapter {
     return encodeSchema(this.schema)
   }
 
-  _encodedMigrations(migrations: SchemaMigrations, fromVersion: SchemaVersion): SQL {
-    const { encodeMigrationSteps } = require('./encodeSchema')
+  _migrationSteps(fromVersion: SchemaVersion): ?(MigrationStep[]) {
     const { stepsForMigration } = require('../../Schema/migrations/helpers')
-    const migrationSteps = stepsForMigration({
+    const { migrations } = this
+    // TODO: Remove this after migrations are shipped
+    if (!migrations) {
+      return null
+    }
+    return stepsForMigration({
       migrations,
       fromVersion,
       toVersion: this.schema.version,
     })
-    return encodeMigrationSteps(migrationSteps)
+  }
+
+  _encodeMigrations(steps: MigrationStep[]): SQL {
+    const { encodeMigrationSteps } = require('./encodeSchema')
+    return encodeMigrationSteps(steps)
   }
 }
